+++ src/main/java/com/atlauncher/LogManager.java
@@ -25,6 +25,7 @@ import java.util.concurrent.BlockingQueue;
 import com.atlauncher.evnt.LogEvent;
 import com.atlauncher.evnt.LogEvent.LogType;
 import com.atlauncher.network.Analytics;
+import com.atlauncher.network.ErrorReporting;
 import com.atlauncher.thread.LoggingThread;
 import com.atlauncher.utils.Utils;
 
@@ -95,6 +96,7 @@ public final class LogManager {
         CharArrayWriter writer = new CharArrayWriter();
         try {
             Analytics.sendException(t.getMessage());
+            ErrorReporting.reportError(t);
             t.printStackTrace(new PrintWriter(writer));
             error(writer.toString());
         } finally {