+++ src/main/java/com/atlauncher/data/InstanceV2Launcher.java
@@ -26,7 +26,10 @@ import com.atlauncher.data.minecraft.loaders.LoaderVersion;
 
 @Json
 public class InstanceV2Launcher {
+    @SuppressWarnings("unused")
     private String launcher = Constants.LAUNCHER_NAME;
+
+    @SuppressWarnings("unused")
     private String launcherVersion = Constants.VERSION.toString();
 
     public String name;