+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 - When installing an instance, give focus to name field and select all when clicked
 - [#255] Add tool to download and use recommended version of Java to use
 - [#350] Add in anonymous analytics to track usage for future decision making
+- Add in Sentry error reporting to get information about unexpected client errors